+++ src/features/homebase/lib/getMaxPersonnel.ts
@@ -1,12 +1,4 @@
-import { PERIODS_TIME, TABLE_MAX_PERSONNEL } from "../model/constants";
-
-export function isPeriodStarted(period: string): boolean {
-  const now = new Date();
-  const [hour, minute] = PERIODS_TIME[period].split(":").map(Number);
-  const periodTime = new Date();
-  periodTime.setHours(hour, minute, 0, 0);
-  return now >= periodTime;
-}
+import { TABLE_MAX_PERSONNEL } from "../model/constants";
 
 export function getMaxPersonnel(
   floor: string,
